Models cleanup

+++ models/message.js
@@ -7,16 +7,19 @@ class Message {
     static async create(data) {
         /** Create a message (from data), update db, return new message data
          *
-         * Data should be: { email, name, message }
+         * Accepts data
+         *      data should be: { email, name, message }
          *
          * Returns { id, email, name, message, received }
          *
          * Throws BadRequestError if incomplete or no data
          */
+        // check for missing / incomplete data
         if (!data) throw new BadRequestError("No data.");
         if (!data.email || !data.name || !data.message)
             throw new BadRequestError("Missing data.");
 
+        // query db to create new message
         const result = await db.query(
             `INSERT INTO messages (email,
                                     name,
@@ -40,8 +43,10 @@ class Message {
          * Throws BadRequestError if no id
          * Throws NotFoundError if no message found
          */
+        // check for missing input
         if (!id) throw new BadRequestError("No id provided.");
 
+        // query db to get message
         const result = await db.query(
             `SELECT id,
                     email,
@@ -55,6 +60,7 @@ class Message {
         );
         const message = result.rows[0];
 
+        // if no record returned, no message found, throw NotFoundError
         if (!message) throw new NotFoundError(`No message: ${id}`);
 
         return message;
@@ -66,6 +72,7 @@ class Message {
          * Returns [{ id, email, name, message, received, isArchived },
          *              { id, email, name, message, received, isArchived }, ...]
          */
+        // query db for list of all messages
         const result = await db.query(
             `SELECT id,
                     email,
@@ -80,11 +87,12 @@ class Message {
     }
 
     static async getActive() {
-        /** Get an array of messages that are NOT marked as archived or deleted
+        /** Get an array of messages that are NOT marked as archived
          *
          * Returns [{ id, email, name, message, received },
          *              { id, email, name, message, received }, ...]
          */
+        // query db for list of all non-archived messages
         const result = await db.query(
             `SELECT id,
                     email,
@@ -92,7 +100,7 @@ class Message {
                     message,
                     received
                 FROM messages
-                WHERE is_archived = false AND is_deleted = false`
+                WHERE is_archived = false`
         );
 
         console.log(JSON.stringify(result.rows));
@@ -106,6 +114,7 @@ class Message {
          * Returns [{ id, email, name, message, received },
          *              { id, email, name, message, received }, ...]
          */
+        // query db for list of all archived messages
         const result = await db.query(
             `SELECT id,
                     email,
@@ -113,8 +122,7 @@ class Message {
                     message,
                     received
                 FROM messages
-                WHERE is_archived = true
-                AND is_deleted = false`
+                WHERE is_archived = true`
         );
 
         return result.rows;
@@ -130,8 +138,10 @@ class Message {
          * Throws BadRequestError if no id
          * Throws NotFoundError if no message found by id
          */
+        // check for missing input
         if (!id) throw new BadRequestError("No id provided.");
 
+        // query db to update message as archived
         const result = await db.query(
             `UPDATE messages
                 SET is_archived = true
@@ -144,9 +154,9 @@ class Message {
                         is_archived AS "isArchived"`,
             [id]
         );
-
         const message = result.rows[0];
 
+        // if no record returned, no message found, throw NotFoundError
         if (!message) throw new NotFoundError(`No message: ${id}`);
 
         return message;
@@ -162,8 +172,10 @@ class Message {
          * Throws BadRequestError if no id
          * Throws NotFound Error if no message found by id
          */
+        // check for missing input
         if (!id) throw new BadRequestError("No id provided.");
 
+        // query db to change is_archived for message to false
         const result = await db.query(
             `UPDATE messages
                 SET is_archived = false
@@ -176,9 +188,9 @@ class Message {
                         is_archived AS "isArchived"`,
             [id]
         );
-
         const message = result.rows[0];
 
+        // if no record returned, no message found, throw NotFoundError
         if (!message) throw new NotFoundError(`No message: ${id}`);
 
         return message;
@@ -194,6 +206,10 @@ class Message {
          * Throws BadRequestError if no id
          * Throws NotFoundError if not found
          */
+        // check for missing input
+        if (!id) throw new BadRequestError("No input.");
+
+        // query db to delete message
         const result = await db.query(
             `DELETE FROM messages
                 WHERE id=$1
@@ -202,6 +218,7 @@ class Message {
         );
         const removed = result.rows[0];
 
+        // if no record returned, message not found, throw NotFoundError
         if (!removed) throw new NotFoundError(`No message: ${id}`);
 
         return { msg: "Deleted." };
